Add files via upload
This folder includes all the files for processing data, training data, and running data for a three-layer neural network. It uses the NREL 2005 Western Wind dataset as training data, and the NREL 2006 Western Wind dataset as test data. Wind speed is a feature, and turbine power output is the algorithm's predicted result. 

Neural network algorithms can learn through a feedback process called backpropagation, which involves comparing the algorithm’s output with the output it was meant to produce. The difference between these values is then used to modify the weights to alter connections in the functions ex4weights.mat, randInitializeWeights.m, and debugInitializeWeights.m.  The cost function computes the error margin at each iteration, the predict function computes and displays the predicted output, and ex4.m initializes variables and calls other functions. The rest of the functions are used to process the data and initialize more values.