Add files via upload
This file contains the code for linear regression. Linear regression uses a linear function to make a prediction, resulting in a straight line that fits the given data. Each point on the line represents the value of an individual prediction, so a line that accurately fits the data represents a more accurate prediction.